{"schema":"decodex/commit/1","summary":"Extract prepared export controller","authority":"manual"}
